Now compiling on Windows 10 with VS2019
Now compiling on Windows 10 with VS2019, some POSIX standard updates i.e. _mkdir instead of mkdir, and some stdlib includes to ensure MALLOC was allocating void* pointers.

+++ doomgeneric/r_main.c
@@ -309,7 +309,7 @@ R_PointToAngle
 	    if (x>y)
 	    {
 		// octant 8
-		return -tantoangle[SlopeDiv(y,x)];
+		return -(int)tantoangle[SlopeDiv(y,x)];
 	    }
 	    else
 	    {
