+++ doomgeneric/r_segs.c
@@ -618,7 +618,7 @@ R_StoreWallRange
 	offsetangle = rw_normalangle-rw_angle1;
 	
 	if (offsetangle > ANG180)
-	    offsetangle = -offsetangle;
+	    offsetangle = -(int)offsetangle;
 
 	if (offsetangle > ANG90)
 	    offsetangle = ANG90;
